@uppy/companion: pass dynamic parameters to grant
This allows to define dynamic properties for grant providers.
E.g. `[subdomain]` can be automatically replaced in `authorize_url` and `access_url`.
For that to work we need to include the dynamic property in the redirect
to the grant middleware. We merge the grantConfig back to the providerConfig,
so we can determine the required query params based on the `dynamic` property.
This avoids an attacker overriding unexpected properties (if we simply
serialize and pass along `query`)